feat: 시도 횟수 입력 기능(2번째) 구현 및 시도 횟수 유효성 검사 추가
- utils/errors/car-name-validator.js 파일 삭제 후 utils/validator.js에서 유효성 검사를 한번에 관리하도록 변경
- 시도 횟수가 숫자인지, 정수인지, 양의 정수인지 확인하는 로직 추가

예시:
입력: abc
에러: [ERROR] 입력값은 숫자여야 합니다.

입력: 3.5
에러: [ERROR] 정수를 입력해주세요.

입력: -1
에러: [ERROR] 시도 횟수는 양의 정수여야 합니다.
